refactor: DetailVC에서 diaryItem 프로퍼티 옵셔널 제거, DiaryModel init 추가

+++ Diary/Etc/DateExtension.swift
@@ -15,8 +15,8 @@ extension Date {
         return formatter
     }()
     
-    static func convertToDate(by date: Int) -> String {
-        let date = Date(timeIntervalSince1970: Double(date))
+    static func convertToDate(by date: Double) -> String {
+        let date = Date(timeIntervalSince1970: date)
         let dateText = Date.formatter.string(from: date)
         
         return dateText
